// AsuguradoraAngular/AseguradoraFront/src/app/Models/Cliente.ts
export interface Cliente{
    idCliente: number;
    documentoIdentidad: string;
    primerNombre: string;
    segundoNombre: string;
    primerApellido: string;
    segundoApellido: string;
    telefono: string;
    email: string;
    fechaNacimiento: string;
    valorSeguro: number;
    observaciones: string;
}